% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{count_responses}
\alias{count_responses}
\title{Calculate the number of patient feedback received by week/ month/ quarter}
\usage{
count_responses(data, period, mode, area = NA)
}
\arguments{
\item{data}{Dataframe, filtered by date/ area in a reactive function}

\item{period}{string, either "week", "month", "year", indicating the preferred
granularity with respect to time}

\item{mode}{boolean- return grouped by mode of response or all together}

\item{area}{level of trust to report at- NA for Trust, or string giving
grouping variable- Division, Directorate, Team}
}
\description{
This function takes a (filtered) dataset and returns the numbers
of pieces of patient feedback received by week/ month/ year. Optionally 
select a split by mode of feedback
}
